# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "gestion_usuarios",
        "description": "Proyecto API REST hecho en Laravel para la gestión de usuarios de una base de datos MySQL",
        "contact": {
            "email": "juanmanuel.pernia@ara-tech.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "http://127.0.0.1:80/gestion_usuarios/public/"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List all users",
                "description": "List all users",
                "operationId": "ListUsers",
                "responses": {
                    "200": {
                        "description": "List of all successful users"
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    },
                    "404": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Store a new user",
                "description": "Create and store a new user in the database",
                "operationId": "be551c1d694a01c164966f58bfa77013",
                "requestBody": {
                    "description": "User to store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "rol_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The name of the user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "The email of the user",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "The password of the user",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "Password confirmation that should match the password",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "rol_id": {
                                        "description": "The role id of the user",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User successfully created"
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    },
                    "422": {
                        "description": "Error: Unprocessable Content"
                    },
                    "500": {
                        "description": "Error: Internal Server Error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show a specific user",
                "description": "Show a specific user",
                "operationId": "6f51bd5fa2dddcd25e44cfa53dcbd10a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the user to show",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found"
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    },
                    "404": {
                        "description": "Error: Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a specific user",
                "description": "Update a specific user in the database",
                "operationId": "070e37fc9aff78ae73860751bc990ed3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the user to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of the user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "The email of the user",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "rol_id": {
                                        "description": "The role id of the user",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User successfully modified"
                    },
                    "404": {
                        "description": "Error: Not Found"
                    },
                    "422": {
                        "description": "Error: Unprocessable Content"
                    },
                    "500": {
                        "description": "Error: Internal Server Error"
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a specific user",
                "description": "Delete a specific user",
                "operationId": "0b4133ed462528a5e648b88ead11e929",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully eliminated "
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    },
                    "404": {
                        "description": "Error: Not found"
                    },
                    "500": {
                        "description": "Error: Internal Server Error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "API Key Auth",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}